pyproject.toml
[project]
name = "prjdjango"
version = "0.1.0"
description = ""
authors = [{ name = "wizli595", email = "abdoowizli@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["django (>=5.2,<6.0)", "pillow (>=11.2.1,<12.0.0)", "django-allauth[socialaccount] (>=65.7.0,<66.0.0)", "django-widget-tweaks (>=1.5.0,<2.0.0)", "django-browser-reload (>=1.18.0,<2.0.0)", "django-htmx (>=1.23.0,<2.0.0)", "langchain-openai (>=0.3.16,<0.4.0)", "langchain-core (>=0.3.59,<0.4.0)", "python-dotenv (>=1.1.0,<2.0.0)"]

[tool.poetry]
packages = [{ include = "prjdjango" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
